Move custom chat components to individual files, took a first pass at displaying image analysis results
