+++ src/pages/Chat.js
@@ -2,17 +2,13 @@ import React from 'react';
 import { ThemeProvider } from 'styled-components';
 import ChatBot from 'react-simple-chatbot';
 import { Container, Row, Col } from 'reactstrap';
-import axios from 'axios'
 import 'rc-time-picker/assets/index.css';
 import moment from 'moment';
 import TimePicker from 'rc-time-picker';
-import { Redirect } from 'react-router-dom'
-import GoogleMapsLoader from 'google-maps'
-
-// Setting Google Maps library variables
-GoogleMapsLoader.KEY = process.env.REACT_APP_GOOGLE_KEY;
-GoogleMapsLoader.LIBRARIES = ['places'];
-GoogleMapsLoader.LANGUAGE = 'en';
+import { Redirect } from 'react-router-dom';
+import ItinerarySearch from '../containers/ItinerarySearch';
+import ImageReview from '../containers/ImageReview';
+import ItineraryConfirm from '../containers/ItineraryConfirm';
 
 // Formatting date/time to be saved in local storage
 const format = 'YYYY-MM-DD HH:mm';
@@ -38,13 +34,14 @@ export default class Chat extends React.Component {
         loading: true,
         itineraryTime: '',
         call: false,
-        update: true
+        update: true,
     }
 
     componentDidMount = () => {
         this.setState({
             firstName: localStorage.getItem('firstName'),
-            loading: false
+            loading: false,
+            update: true,
         })
         if (!localStorage.getItem('token')) {
             this.setState({
@@ -72,29 +69,34 @@ export default class Chat extends React.Component {
 
     render() {
         const loader = <div>Loading...</div>
-        const update = [{
-            id: '1',
-            message: 'Aunty has an update for you.',
-            trigger: '2',
+        const steps = [{
+            id: 'welcome',
+            message: `Hallo ${this.state.firstName}! Aunty here. How are you?`,
+            trigger: () => this.state.update ? 'update' : 'welcome user response'
+        },{
+            id: 'update',
+            component: (
+                <ImageReview />
+            ),
+            asMessage: true,
+            trigger: 'need more help'
         }, {
-            id: '2',
-            message: 'I took a look at your photo.',
-            trigger: '3'
+            id: 'need more help',
+            message: 'Need anything else?',
+            trigger: 'need more help options'
+
         }, {
-            id: '3',
-            message: 'This is what I found...',
-            end: true
-        }];
-        const normal = [{
-            id: '1',
-            message: `Hallo hallo ${this.state.firstName}! Aunty here. How are you?`,
-            trigger: '2',
-        }, {
-            id: '2',
+            id: 'need more help options',
+            options: [
+            { value: 'Yes', label: 'Yes', trigger: 'menu' },
+            { value: 'No', label: 'No', trigger: '21' },
+            ],
+      },{
+            id: 'welcome user response',
             user: true,
-            trigger: '3',
+            trigger: 'menu',
         }, {
-            id: '3',
+            id: 'menu',
             message: 'How can Aunty help you today?',
             trigger: '4',
         }, {
@@ -131,7 +133,7 @@ export default class Chat extends React.Component {
         }, {
             id: 'itineraryLocationSearch',
             component: (
-                <SearchItineraryLocation />
+                <ItinerarySearch />
             ),
             waitAction: true,
             asMessage: true,
@@ -217,7 +219,7 @@ export default class Chat extends React.Component {
         }, {
             id: 'ConfirmItineraryPin',
             component: (
-                <ConfirmItineraryPin />
+                <ItineraryConfirm />
             ),
             asMessage: true,
             trigger: '19'
@@ -262,153 +264,22 @@ export default class Chat extends React.Component {
         }];
         return (
             <Container>
-                <Row className="justify-content-md-center">
-                    <Col md="auto">
-                    {
-                        this.state.loading || this.state.firstName.length === 0 ?
-                        loader 
-                        :
-                        <ThemeProvider theme={theme}>
-                            <ChatBot
-                                cache={true}
-                                hideHeader={true}
-                                headerTitle="Chat with Aunty"
-                                recognitionEnable={true}
-                                speechSynthesis={{ enable: true, lang: 'en' }}
-                                steps={this.state.update ? update : normal}
-                            />
-                        </ThemeProvider>
-                    }
-                    </Col>
-                </Row>
-                { this.handleRedirect() }
+                {
+                    this.state.loading || this.state.firstName.length === 0 ?
+                    loader 
+                    :
+                    <ThemeProvider theme={theme}>
+                        <ChatBot
+                            hideHeader={true}
+                            headerTitle="Chat with Aunty"
+                            recognitionEnable={true}
+                            // speechSynthesis={{ enable: true, lang: 'en' }}
+                            steps={steps}
+                        />
+                    </ThemeProvider>
+                }
+            { this.handleRedirect() }
             </Container>
         )
-    }
-}
-
-class ConfirmItineraryPin extends React.Component {
-    state = {
-        location: '',
-        description: '',
-        time: ''
-    };
-        
-    componentWillMount() {
-        const userId = localStorage.getItem('userId');
-        const location = localStorage.getItem('itineraryLocation');
-        const time = localStorage.getItem('itineraryTime');
-        const description = localStorage.getItem('itineraryDetails');
-        const latitude = localStorage.getItem('itineraryLatitude');
-        const longitude = localStorage.getItem('itineraryLongitude');
-
-        // Make axios POST request to create itinerary pin 
-        axios({
-            method: 'POST',
-            url: 'https://gokaikai.herokuapp.com/api/v1/pins/itinerary/',
-            data: {
-                pinName: location,
-                userId: userId,
-                longitude: parseFloat(latitude),
-                latitude: parseFloat(longitude),
-                startTime: time,
-                description: description
-            }
-        })
-        .then(response => {
-            console.log(response.data);
-        })
-        .catch(error => {
-            console.log(error);
-        }) 
-        this.setState({
-            location: location,
-            description: description,
-            time: time
-        })
-  }
-  render() {
-    return (
-        <div>
-            <p>Ok so you are going to {this.state.location} {this.state.description} at {this.state.time}?</p>
-            <p>Aunty will check in then, lah!</p>
-        </div>
-    );
-  }
-}
-
-class SearchItineraryLocation extends React.Component {
-    state = {
-        error: false,
-        name: '',
-        address: '',
-        loading: true
-    };
-    componentWillMount() {
-
-        // Retrieving user's coords from local storage
-        const userLat = parseFloat(localStorage.getItem('latitude'))
-        const userLong = parseFloat(localStorage.getItem('longitude'))
-
-        // Radius is set to 100 km / 60 miles around user's location
-        const userLocation = {radius: 100000, center: {lat: userLat, lng: userLong}}
-
-        // Request returns location name, coords, and address
-        const request = {
-            query: this.props.steps.itineraryLocation.value,
-            fields: ['name', 'geometry', 'formatted_address'],
-            locationBias: userLocation
-        }
-    
-        GoogleMapsLoader.load((google) => {
-            // Creating / attaching a node required to use API
-            const map = new google.maps.Map(document.createElement('div'));
-            const service = new google.maps.places.PlacesService(map);
-            
-            // Sending query text-based request to find place
-            service.findPlaceFromQuery(request, (results, status) => {
-                if (status === "OK") {
-                    this.props.triggerNextStep({ trigger: '8' })
-                    this.setState({
-                        loading: false,
-                        name: results[0]['name'],
-                        address: results[0]['formatted_address']
-                    })
-                    // Must call functions to find lat/lng
-                    const lat = results[0]['geometry']['location'].lat().toFixed(6)
-                    const lng = results[0]['geometry']['location'].lng().toFixed(6)
-                    localStorage.setItem('itineraryLatitude', lat)
-                    localStorage.setItem('itineraryLongitude', lng)
-                    localStorage.setItem('itineraryLocation', this.state.name);
-                } else {
-                    this.props.triggerNextStep({
-                        trigger: '5'
-                    })
-                    this.setState({
-                        error: true,
-                        loading: false,
-                    })
-                }
-            })
-        });
-    }
-
-    render() {
-        const { name, address, error, loading } = this.state;
-        return (
-            <div> {
-                loading 
-                ? <p>Searching my phonebook...</p>
-                : error 
-                    ? <p>
-                        I don't know where that is! Please try again.
-                    </p>
-                    : <div id="result">
-                        <p>Is it this one ah?</p>
-                        <p><b>{ name }</b></p>
-                        <p>{ address }</p>
-                </div> }
-            </div>
-        );
     }
 }
